+++ packages/yarnpkg-core/sources/Project.ts
@@ -531,20 +531,6 @@ export class Project {
     return workspace;
   }
 
-  /**
-   * Import the dependencies of each resolved workspace into their own
-   * `Workspace` instance.
-   */
-  private refreshWorkspaceDependencies() {
-    for (const workspace of this.workspaces) {
-      const pkg = this.storedPackages.get(workspace.anchoredLocator.locatorHash);
-      if (!pkg)
-        throw new Error(`Assertion failed: Expected workspace ${structUtils.prettyWorkspace(this.configuration, workspace)} (${formatUtils.pretty(this.configuration, ppath.join(workspace.cwd, Filename.manifest), formatUtils.Type.PATH)}) to have been resolved. Run "yarn install" to update the lockfile`);
-
-      workspace.dependencies = new Map(pkg.dependencies);
-    }
-  }
-
   forgetResolution(descriptor: Descriptor): void;
   forgetResolution(locator: Locator): void;
   forgetResolution(dataStructure: Descriptor | Locator): void {
@@ -943,11 +929,6 @@ export class Project {
     this.originalPackages = originalPackages;
     this.optionalBuilds = optionalBuilds;
     this.peerRequirements = peerRequirements;
-
-    // Now that the internal resolutions have been updated, we can refresh the
-    // dependencies of each resolved workspace's `Workspace` instance.
-
-    this.refreshWorkspaceDependencies();
   }
 
   async fetchEverything({cache, report, fetcher: userFetcher, mode}: InstallOptions) {
@@ -1808,7 +1789,6 @@ export class Project {
     if (restoreResolutions) {
       if (installState.lockFileChecksum === this.lockFileChecksum) {
         Object.assign(this, pick(installState, INSTALL_STATE_FIELDS.restoreResolutions));
-        this.refreshWorkspaceDependencies();
       } else {
         await this.applyLightResolution();
       }